pyproject.toml
[tool.poetry]
name = "fast-template"
version = "0.1.0"
description = ""
authors = ["sunling <arniss472@gmail.com>"]
readme = "README.md"
packages = [{include = "fast-template"}]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.20.0"
fastapi = "^0.92.0"
pydantic = "^1.10.5"
asyncpg = "^0.27.0"
sqlalchemy = "^2.0.4"
alembic = "^1.9.4"
environs = "^9.5.0"
py-bcrypt = "^0.4"
python-multipart = "^0.0.6"
pytest = "^7.3.1"
httpx = "^0.24.0"
rocketry = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
